phenomedb/requirements.txt
alembic==1.4.3
apache-airflow==2.0.0
apache-airflow-providers-ftp==1.0.0
apache-airflow-providers-http==1.0.0
apache-airflow-providers-imap==1.0.0
apache-airflow-providers-sqlite==1.0.0
apache-airflow-upgrade-check==1.1.0
apispec==3.3.2
appdirs==1.4.4
appnope==0.1.0
argcomplete==1.12.2
asteval==0.9.19
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.1
biopython==1.78
cached-property==1.5.1
cattrs==1.1.2
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
ChemSpiPy==2.0.0
click==7.1.2
clickclick==20.10.2
colorama==0.4.4
colorlog==4.0.2
commonmark==0.9.1
configparser==3.5.3
connexion==2.7.0
croniter==0.3.36
cryptography==3.3.1
cycler==0.10.0
decorator==4.4.2
deepdiff==5.0.2
defusedxml==0.6.0
dill==0.3.3
dnspython==2.0.0
docutils==0.16
email-validator==1.1.2
et-xmlfile==1.0.1
Flask==1.1.2
Flask-Admin==1.5.4
Flask-AppBuilder==3.1.1
Flask-Babel==1.0.0
Flask-Bcrypt==0.7.1
Flask-Caching==1.9.0
Flask-JWT-Extended==3.25.0
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.4
flask-swagger==0.2.13
Flask-WTF==0.14.3
fs==2.4.11
funcsigs==1.0.2
future==0.18.2
graphviz==0.16
gunicorn==19.10.0
idna==2.10
importlib-metadata==1.7.0
importlib-resources==1.5.0
inflection==0.5.1
ipython==7.18.1
ipython-genutils==0.2.0
isaExplorer==0.1.0
isatools==0.11.0
iso8601==0.1.13
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
joblib==0.16.0
json-merge-patch==0.2
jsonschema==3.2.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lmfit==1.0.1
lockfile==0.12.2
lxml==4.5.2
Mako==1.1.3
Markdown==3.3.3
MarkupSafe==1.1.1
marshmallow==3.10.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==2.1.0
marshmallow-sqlalchemy==0.23.1
matplotlib==3.3.1
mzml2isa==1.0.3
natsort==7.1.0
networkx==2.5
nPYc==1.2.3
numpy==1.19.2
openapi-spec-validator==0.2.9
openpyxl==2.6.4
ordered-set==4.0.2
packaging==20.4
pandas==1.1.2
parso==0.7.1
patsy==0.5.1
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pip-autoremove==0.9.1
plotly==4.10.0
prison==0.1.3
progressbar2==3.53.1
prompt-toolkit==3.0.7
pronto==0.12.2
psutil==5.8.0
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyChemometrics==0.1
pycparser==2.20
Pygments==2.7.0
PyJWT
pyparsing==2.4.7
pyrsistent==0.17.3
python-daemon==2.2.4
python-dateutil==2.8.1
python-editor==1.0.4
python-ldap==3.3.1
python-nvd3==0.15.0
python-slugify==4.0.1
python-utils==2.4.0
python3-openid==3.2.0
pytz==2020.1
pytzdata==2020.1
PyYAML==5.3.1
requests==2.23.0
retrying==1.3.3
rich==9.2.0
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.11.0
setproctitle==1.2.1
simplejson==3.17.2
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.22
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.36.8
statsmodels==0.12.0
swagger-ui-bundle==0.0.8
tabulate==0.8.7
tenacity==6.2.0
termcolor==1.1.0
text-unidecode==1.3
threadpoolctl==2.1.0
thrift==0.13.0
traitlets==5.0.4
typing-extensions==3.7.4.3
tzlocal==1.5.1
uncertainties==3.1.4
unicodecsv==0.14.1
urllib3==1.25.10
wcwidth==0.2.5
Werkzeug==1.0.1
WTForms==2.3.3
xlrd==1.2.0
zipp==3.4.0
zope.deprecation==4.4.0
